refactor: 파일 관련 예외 한 폴더로 모음
