+++ src/main/java/com/example/matchup/matchupbackend/service/FileService.java
@@ -5,7 +5,7 @@ import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.example.matchup.matchupbackend.dto.UploadFile;
-import com.example.matchup.matchupbackend.error.exception.FileUploadException;
+import com.example.matchup.matchupbackend.error.exception.FileEx.FileUploadException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
